20150218-5 Cleanup and Scripts
All character behavior is now in obj_character.  There are two new
script groups that pertain to characters, one is for character death
behavior and the other is to get input or run AI.  I've also made it so
that player death resets the game and enemy death results in enemies
removing themselves from the enemy list that holds them all after
created in the room creation code.